refactored the states
