+++ frontend/src/App.jsx
@@ -10,6 +10,7 @@ import {
   QueryClientProvider
 } from '@tanstack/react-query'
 import MainPage from './components/MainPage.jsx';
+import BookSearchProvider, { BookSearchContext } from './context/BookSearchContext.jsx';
 
 const queryClient = new QueryClient()
 
@@ -17,16 +18,19 @@ const App = () => {
   return (
     <>
       <QueryClientProvider client={queryClient}>
-        <BrowserRouter>
-          <Routes>
-            <Route path='search' element={<SearchPage />}>
-              <Route index element={<MainPage />}/>
-              <Route path='title' element={<TitlePageResults />}></Route>
-              <Route path='author'></Route>
-              <Route path='title-author'></Route>
-            </Route>
-          </Routes>
-        </BrowserRouter>
+        <BookSearchProvider>  
+          <BrowserRouter>
+            <Routes>
+              
+                <Route path='search' element={<SearchPage />}>
+                  <Route index element={<MainPage />}/>
+                  <Route path='title' element={<TitlePageResults />}></Route>
+                  <Route path='author'></Route>
+                  <Route path='title-author'></Route>
+                </Route>
+            </Routes>
+          </BrowserRouter>
+        </BookSearchProvider>
       </QueryClientProvider>
     </>
   )
