+++ frontend/src/components/ui/Quotes.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import Loading from "../Loading";
 
 const Quotes = ({ quoteData }) => {
@@ -9,9 +9,19 @@ const Quotes = ({ quoteData }) => {
   // } else {
   //   console.log("Not exist yet")
   // }
-  const {q , a} = quoteData[0];
-    
-  return (
+  const [quote , setQuote] = useState(null);
+
+  useEffect(() => {
+  if (quoteData) {
+      setQuote(quoteData[0]); // one quote object
+    }
+  }, [quoteData]);
+
+  if (!quote) return <Loading />;
+  const { q, a } = quote;
+
+
+  return quote && (
     <div className="quotes-main flex w-full h-full">
       <div className="quotes-wrapper flex justify-center items-center w-full">
         <div className="quotes-inner bg-zinc-900 w-[565px] h-[265px] rounded-2xl border-1 border-zinc-600 flex gap-2">
